// legacy-structure/client/src/modules/finance/models/models.ts
export interface CollaboratorRecord {
  name: string;
  avatar_url: string;
  card_number: string;
  id: number;
  login: string;
  node_id: string;
  pay_by_hours: number;
  // pay_to_internet: number;
  url: string;
  email: string;
  address: string;
  cap: string;
  place: string;
  bio: string;
  isCollaborator: boolean;
  type: string;
  typeTranslation: string;
}
// ...
export interface MediaListRow {
  id: number;
  fileName?: string;
  type?: string;
  displayName?: string;
}
// ...
export interface ProductRecord {
  id: number;
  name: string;
  description: string;
  buyPrice: number;
  sellPrice: number;
  unitCount: number;
  type: number;
  docs: MediaListRow[] | number[];
}
// ...
export interface ProductsBill {
  productId: number;
  product: ProductRecord;
  price: number;
  count: number;
  maxCount: number;
  remove: boolean;
  type: number;
}

export interface ColumnProductsBillTable extends IColumn {
  field?: (row: ProductsBill) => string;
}

export interface BillRecord {
  id: number;
  name: string;
  description: string;
  totalPrice: number;
  totalWithDiscount: number;
  clientId: number;
  date: Date;
  sendDate: Date;
  discount: number;
  observations: string;
  status: number;
  validDays: number;
  client: CollaboratorRecord;
  products: ProductsBill[];
  docs: MediaListRow[] | number[];
}

export interface BillFilter {
  status?: { value: number; label: string };
  client: string;
}

export type BillKeys =
  | 'id'
  | 'name'
  | 'description'
  | 'totalPrice'
  | 'clientId';

export interface ColumnBillTable extends IColumn {
  field?: ((row: BillRecord) => string) | BillKeys;
}

export interface BillDataResponse {
  bills?: BillRecord[];
  updated?: boolean;
  created?: boolean;
}
export interface BillResponse {
  data: BillDataResponse;
}

export interface TemplateStyle {
  body: string;
  defaultTemplate: boolean;
  documentType: string;
  entityTemplate: string;
  id: number;
  title: string;
}

export interface TemplateHTML {
  body: string;
  defaultTemplate: boolean;
  documentType: string;
  entityTemplate: string;
  id: number;
  styleTemplate: TemplateStyle;
  styleTemplateId: number;
  title: string;
}